re: Don't compare literal strings one character at a time
Introduce a new opcode, LITERAL_STRING, that matches a whole string
rather than each individual character of a string as it happens with
the LITERAL opcode.  This reduces the instruction dispatch overhead
within the _sre module.

Instead of storing a pointer to a string inline with the code, the
argument to LITERAL_STRING is an index to a new "literals" list.

The overlap table and the prefix isn't necessary, so they're
not generated anymore.